requirements.txt
# Core Framework Dependencies
click>=8.1.7
pydantic>=2.4.2
loguru>=0.7.2

# Document Processing
unstructured[all-docs]>=0.10.30
pymupdf>=1.23.8
pytesseract>=0.3.10
python-docx>=0.8.11
python-pptx>=0.6.21
html2text>=2020.1.16

# Image Processing
opencv-python>=4.8.0
pillow>=10.0.1
scikit-image>=0.21.0

# Audio Processing
faster-whisper>=0.9.0
ffmpeg-python>=0.2.0

# LLM Integration
ollama>=0.1.7
transformers>=4.35.0

# System Utilities
psutil>=5.9.6
numpy>=1.24.3
requests>=2.31.0

# Development Dependencies
pytest>=7.4.3
pytest-cov>=4.1.0
black>=23.9.1
flake8>=6.1.0
mypy>=1.6.1
